pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepseek-ocr2-batch-rag"
version = "0.1.0"
description = "Non-invasive PDF batch extension for DeepSeek-OCR2 focused on RAG preprocessing."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Maintainers" }]
dependencies = [
  "vllm",
  "transformers",
  "PyMuPDF",
  "img2pdf",
  "numpy",
  "Pillow",
  "tqdm",
  "PyYAML",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
deepseek-ocr2-batch-rag = "main:main"

[tool.setuptools]
py-modules = ["main", "config_loader", "gpu_manager", "pathing", "pipeline", "scanner", "worker"]
